+++ src/test/java/de/tum/in/www1/artemis/programmingexercise/ProgrammingExerciseTestCaseServiceTest.java
@@ -26,6 +26,7 @@ import de.tum.in.www1.artemis.domain.hestia.ProgrammingExerciseTestCaseType;
 import de.tum.in.www1.artemis.repository.ProgrammingExerciseRepository;
 import de.tum.in.www1.artemis.repository.ProgrammingExerciseTestCaseRepository;
 import de.tum.in.www1.artemis.security.SecurityUtils;
+import de.tum.in.www1.artemis.service.programming.ProgrammingExerciseFeedbackService;
 import de.tum.in.www1.artemis.service.programming.ProgrammingExerciseTestCaseService;
 import de.tum.in.www1.artemis.util.ModelFactory;
 import de.tum.in.www1.artemis.web.rest.dto.ProgrammingExerciseTestCaseDTO;
@@ -39,6 +40,9 @@ class ProgrammingExerciseTestCaseServiceTest extends AbstractSpringIntegrationBa
 
     @Autowired
     private ProgrammingExerciseTestCaseService testCaseService;
+
+    @Autowired
+    private ProgrammingExerciseFeedbackService programmingExerciseFeedbackService;
 
     @Autowired
     private ProgrammingExerciseRepository programmingExerciseRepository;
@@ -64,7 +68,7 @@ class ProgrammingExerciseTestCaseServiceTest extends AbstractSpringIntegrationBa
     @Test
     void shouldSetAllTestCasesToInactiveIfFeedbackListIsEmpty() {
         List<Feedback> feedbacks = new ArrayList<>();
-        testCaseService.generateTestCasesFromFeedbacks(feedbacks, programmingExercise);
+        programmingExerciseFeedbackService.generateTestCasesFromFeedbacks(feedbacks, programmingExercise);
 
         Set<ProgrammingExerciseTestCase> testCases = testCaseRepository.findByExerciseId(programmingExercise.getId());
         assertThat(testCases).hasSize(3);
@@ -79,7 +83,7 @@ class ProgrammingExerciseTestCaseServiceTest extends AbstractSpringIntegrationBa
         feedbacks.add(new Feedback().text("test2"));
         feedbacks.add(new Feedback().text("test4"));
         feedbacks.add(new Feedback().text("test5"));
-        testCaseService.generateTestCasesFromFeedbacks(feedbacks, programmingExercise);
+        programmingExerciseFeedbackService.generateTestCasesFromFeedbacks(feedbacks, programmingExercise);
 
         Set<ProgrammingExerciseTestCase> testCases = testCaseRepository.findByExerciseId(programmingExercise.getId());
         assertThat(testCases).hasSize(5);
@@ -102,7 +106,7 @@ class ProgrammingExerciseTestCaseServiceTest extends AbstractSpringIntegrationBa
         List<Feedback> feedbacks = new ArrayList<>();
         feedbacks.add(new Feedback().text("test1"));
         feedbacks.add(new Feedback().text("test2"));
-        testCaseService.generateTestCasesFromFeedbacks(feedbacks, programmingExercise);
+        programmingExerciseFeedbackService.generateTestCasesFromFeedbacks(feedbacks, programmingExercise);
 
         Set<ProgrammingExerciseTestCase> testCases = testCaseRepository.findByExerciseId(programmingExercise.getId());
         assertThat(testCases).hasSize(2);
@@ -116,7 +120,7 @@ class ProgrammingExerciseTestCaseServiceTest extends AbstractSpringIntegrationBa
         testCaseRepository.deleteAll(testCaseRepository.findByExerciseId(programmingExercise.getId()));
 
         List<Feedback> feedbackList = ModelFactory.generateStaticCodeAnalysisFeedbackList(5);
-        testCaseService.generateTestCasesFromFeedbacks(feedbackList, programmingExercise);
+        programmingExerciseFeedbackService.generateTestCasesFromFeedbacks(feedbackList, programmingExercise);
 
         Set<ProgrammingExerciseTestCase> testCases = testCaseRepository.findByExerciseId(programmingExercise.getId());
         assertThat(testCases).isEmpty();
@@ -132,7 +136,7 @@ class ProgrammingExerciseTestCaseServiceTest extends AbstractSpringIntegrationBa
         feedbacks.add(new Feedback().text("generateTestsForAllClasses"));
         feedbacks.add(new Feedback().text("generateTestsForAllClasses"));
         feedbacks.add(new Feedback().text("generateTestsForAllClasses"));
-        testCaseService.generateTestCasesFromFeedbacks(feedbacks, programmingExercise);
+        programmingExerciseFeedbackService.generateTestCasesFromFeedbacks(feedbacks, programmingExercise);
 
         Set<ProgrammingExerciseTestCase> testCases = testCaseRepository.findByExerciseId(programmingExercise.getId());
         assertThat(testCases).hasSize(2);
@@ -212,7 +216,7 @@ class ProgrammingExerciseTestCaseServiceTest extends AbstractSpringIntegrationBa
         feedbacks.add(new Feedback().text("test1"));
         feedbacks.add(new Feedback().text("test2"));
         feedbacks.add(new Feedback().text("test3"));
-        testCaseService.generateTestCasesFromFeedbacks(feedbacks, programmingExercise);
+        programmingExerciseFeedbackService.generateTestCasesFromFeedbacks(feedbacks, programmingExercise);
 
         Set<ProgrammingExerciseTestCase> testCases = testCaseRepository.findByExerciseId(programmingExercise.getId());
         Set<ProgrammingExerciseTestCaseDTO> testCaseDTOs = testCases.stream().map(testCase -> {
@@ -236,7 +240,7 @@ class ProgrammingExerciseTestCaseServiceTest extends AbstractSpringIntegrationBa
                 new ProgrammingExerciseTestCase().testName("testMethods[Context]").exercise(programmingExercise),
                 new ProgrammingExerciseTestCase().testName("testAttributes[Starter]").exercise(programmingExercise));
 
-        testCaseService.setTestCaseType(structuralTestCases, ProgrammingLanguage.JAVA);
+        testCaseRepository.setTestCaseType(structuralTestCases, ProgrammingLanguage.JAVA);
         assertThat(structuralTestCases).allMatch(testCase -> testCase.getType() == ProgrammingExerciseTestCaseType.STRUCTURAL);
     }
 
@@ -247,7 +251,7 @@ class ProgrammingExerciseTestCaseServiceTest extends AbstractSpringIntegrationBa
                 new ProgrammingExerciseTestCase().testName("test13412").exercise(programmingExercise),
                 new ProgrammingExerciseTestCase().testName("HiddenRandomTest").exercise(programmingExercise));
 
-        testCaseService.setTestCaseType(behavioralTestCases, ProgrammingLanguage.JAVA);
+        testCaseRepository.setTestCaseType(behavioralTestCases, ProgrammingLanguage.JAVA);
         assertThat(behavioralTestCases).allMatch(testCase -> testCase.getType() == ProgrammingExerciseTestCaseType.BEHAVIORAL);
     }
 
@@ -263,7 +267,7 @@ class ProgrammingExerciseTestCaseServiceTest extends AbstractSpringIntegrationBa
                     new ProgrammingExerciseTestCase().testName("testMergeSort").exercise(programmingExercise),
                     new ProgrammingExerciseTestCase().testName("test13412").exercise(programmingExercise),
                     new ProgrammingExerciseTestCase().testName("HiddenRandomTest").exercise(programmingExercise));
-            testCaseService.setTestCaseType(testCases, language);
+            testCaseRepository.setTestCaseType(testCases, language);
             assertThat(testCases).allMatch(testCase -> testCase.getType() == ProgrammingExerciseTestCaseType.DEFAULT);
         }
     }
